[refactor] rationalize lib/util and lib/services
