+++ lib/api/documentExtractor.js
@@ -20,7 +20,8 @@
  */
 
 'use strict';
-const assertionError = require('./errors').wrap('api', 'assert');
+
+const assertionError = require('../util/errors').wrap('api', 'assert');
 const extractors = ([
   {
     methods: {
