+++ lib/api/funnel.js
@@ -38,7 +38,7 @@ const
   errorsManager = require('../util/errors'),
   { errors: { KuzzleError } } = require('kuzzle-common-objects'),
   documentEventAliases = require('../config/documentEventAliases'),
-  DocumentExtractor = require('../util/DocumentExtractor'),
+  DocumentExtractor = require('./documentExtractor'),
   sdkCompatibility = require('../config/sdkCompatibility'),
   RateLimiter = require('./rateLimiter');
