+++ lib/core/cache/cacheEngine.js
@@ -21,9 +21,8 @@
 
 'use strict';
 
-const
-  Bluebird = require('bluebird'),
-  Redis = require('../../services/redis');
+const Bluebird = require('bluebird');
+const Redis = require('../../services/cache/redis');
 
 class CacheEngine {
   constructor (kuzzle) {
