+++ lib/core/plugins/context.js
@@ -28,7 +28,7 @@ const { Kuzzle: KuzzleSDK } = require('kuzzle-sdk');
 const FunnelProtocol = require('../sdk/funnelProtocol');
 const PluginRepository = require('./pluginRepository');
 const IndexStorage = require('../storage/indexStorage');
-const Elasticsearch = require('../../services/elasticsearch');
+const Elasticsearch = require('../../services/storage/elasticsearch');
 const {
   errors,
   models: { RequestContext, RequestInput },
