+++ lib/core/storage/storageEngine.js
@@ -21,7 +21,7 @@
 
 'use strict';
 
-const Elasticsearch = require('../../services/elasticsearch');
+const Elasticsearch = require('../../services/storage/elasticsearch');
 const errorsManager = require('../../util/errors').wrap('services', 'storage');
 const ClientAdapter = require('./clientAdapter');
 const BaseModel = require('./models/baseModel');
