+++ lib/services/cache/redis.js
@@ -21,12 +21,11 @@
 
 'use strict';
 
-const
-  _ = require('lodash'),
-  Bluebird = require('bluebird'),
-  IORedis = require('ioredis'),
-  errorsManager = require('../util/errors').wrap('services', 'cache'),
-  Service = require('./service');
+const _ = require('lodash');
+const Bluebird = require('bluebird');
+const IORedis = require('ioredis');
+const errorsManager = require('../../util/errors').wrap('services', 'cache');
+const Service = require('../service');
 
 /**
  * @class Redis
