+++ lib/services/service.js
@@ -21,9 +21,8 @@
 
 'use strict';
 
-const
-  Bluebird = require('bluebird'),
-  errorsManager = require('../util/errors');
+const Bluebird = require('bluebird');
+const errorsManager = require('../util/errors');
 
 /**
  * Services base class
