+++ lib/services/storage/elasticsearch.js
@@ -28,13 +28,13 @@ const ms = require('ms');
 const Bluebird = require('bluebird');
 const semver = require('semver');
 
-const debug = require('../util/debug')('kuzzle:services:elasticsearch');
-const ESWrapper = require('../util/esWrapper');
-const didYouMean = require('../util/didYouMean');
-const Service = require('./service');
-const { assertIsObject } = require('../util/requestAssertions');
-const errorsManager = require('../util/errors').wrap('services', 'storage');
-const { isPlainObject } = require('../util/safeObject');
+const debug = require('../../util/debug')('kuzzle:services:elasticsearch');
+const ESWrapper = require('./esWrapper');
+const didYouMean = require('../../util/didYouMean');
+const Service = require('../service');
+const { assertIsObject } = require('../../util/requestAssertions');
+const errorsManager = require('../../util/errors').wrap('services', 'storage');
+const { isPlainObject } = require('../../util/safeObject');
 
 const SCROLL_CACHE_PREFIX = '_docscroll_';
