+++ lib/services/storage/esWrapper.js
@@ -23,13 +23,12 @@
 
 'use strict';
 
-const
-  Bluebird = require('bluebird'),
-  _ = require('lodash'),
-  debug = require('./debug')('kuzzle:util:ESCommon'),
-  es = require('@elastic/elasticsearch'),
-  errorsManager = require('./errors').wrap('services', 'storage'),
-  { errors: { KuzzleError } } = require('kuzzle-common-objects');
+const Bluebird = require('bluebird');
+const _ = require('lodash');
+const debug = require('../../util/debug')('kuzzle:services:storage:ESCommon');
+const es = require('@elastic/elasticsearch');
+const errorsManager = require('../../util/errors').wrap('services', 'storage');
+const { errors: { KuzzleError } } = require('kuzzle-common-objects');
 
 const errorMessagesMapping = [
   {
